Finish refactoring editor views into separate files
