+++ OpenShow/slides/editor/views/index.py
