+++ OpenShow/slides/editor/views/segment.py
@@ -1,0 +1,14 @@
+from django.views.generic import CreateView, UpdateView
+from slides.models import Segment
+
+
+class SegmentCreateView(CreateView):
+    model = Segment
+    fields = ['show', 'name', 'order']
+    template_name = 'editor/new_segment_form.html'
+
+
+class SegmentUpdateView(UpdateView):
+    model = Segment
+    fields = ['name', 'order', 'included_deck']
+    template_name = 'editor/edit_segment_form.html'
