+++ OpenShow/slides/editor/views/show.py
@@ -1,6 +1,8 @@
-from django.views.generic import DetailView, CreateView, DeleteView
+from django.views.generic import DetailView, CreateView, DeleteView, UpdateView
 from django.urls import reverse_lazy
+from django.shortcuts import render
 from slides.models import Show
+from slides.editor.forms import SetThemeForm
 
 class ShowEditorView(DetailView):
     queryset = Show.objects.all()
@@ -25,3 +27,22 @@ class ShowDeleteView(DeleteView):
     extra_context = {
         'action': 'delete-show',
     }
+
+
+class SetThemeView(UpdateView):
+    model = Show
+    form_class = SetThemeForm
+    template_name = 'editor/set_theme.html'
+
+
+def check_theme_compatibility(request):
+    form = SetThemeForm(request.POST)
+    if form.is_valid():
+        show = Show.objects.get(pk=form.cleaned_data['show_pk'])
+        theme = form.cleaned_data['theme']
+        missing = show.check_compatibility(theme)
+        if len(missing) == 0:
+            missing = None
+        return render(request, 'editor/show_compatibility_snippet.html', {
+            'missing': missing,
+        })
