+++ OpenShow/slides/editor/views/transition.py
@@ -1,0 +1,46 @@
+from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView
+from slides.models import Transition, TransitionKeyframe
+
+
+class TransitionEditorIndexView(ListView):
+    queryset = Transition.objects.all()
+    template_name = 'editor/transition_editor.html'
+    extra_context = {
+        'transition_list': Transition.objects.all()
+    }
+
+
+class TransitionEditorView(DetailView):
+    model = Transition
+    template_name = 'editor/transition_editor.html'
+    extra_context = {
+        'transition_list': Transition.objects.all
+    }
+
+
+class TransitionCreateView(CreateView):
+    model = Transition
+    fields = ["name"]
+    template_name = 'editor/snippets/hx-simple_create_form.html'
+    extra_context = {
+        'action': 'new-transition',
+        'object_type': 'Transition',
+    }
+
+
+class TransitionKeyframeCreateView(CreateView):
+    model = TransitionKeyframe
+    fields = ["transition", "marker", "css", "out"]
+    template_name = "editor/edit_keyframe.html"
+    extra_context = {
+        'action': 'new-keyframe',
+    }
+
+
+class TransitionKeyframeUpdateView(UpdateView):
+    model = TransitionKeyframe
+    fields = ["marker", "css"]
+    template_name = "editor/edit_keyframe.html"
+    extra_context = {
+        'action': 'edit-keyframe',
+    }
